backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.10.2
anyio==3.6.2
async-timeout==4.0.2
asyncpg==0.27.0
attrs==23.1.0
bcrypt==4.0.1
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.1
Deprecated==1.2.13
distlib==0.3.6
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
fastapi==0.95.0
filelock==3.12.0
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
identify==2.5.22
idna==3.4
Jinja2==3.1.2
jwcrypto==1.4.2
Mako==1.2.4
MarkupSafe==2.1.2
multidict==6.0.4
nodeenv==1.7.0
passlib==1.7.4
platformdirs==3.2.0
pre-commit==3.2.2
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.7
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0
redis==4.5.4
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.8
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.21.1
virtualenv==20.21.0
wrapt==1.15.0
yarl==1.8.2
